Update index.html
I like the look of the page especially the background picture. It may be easier to navigate and attract more users if the site is for a more focused crowd like people who want to play the violin again.  That way you have more of an idea of their skill set.
Another idea for the quiz page is to offer a self assessment because people who have already played will be able to assess themselves.